+++ packages/ui/src/components/common/TransactionModal/steps/DepositAddressValidatingStep.tsx
@@ -1,11 +1,8 @@
 import type { FC } from 'react'
 import { Anchor, Button, Flex, Text } from '../../../primitives/index.js'
 import { LoadingSpinner } from '../../LoadingSpinner.js'
-import type { ExecuteStep, ExecuteStepItem } from '@relayprotocol/relay-sdk'
 import { useExecutionStatus } from '@relayprotocol/relay-kit-hooks'
 import { useRelayClient } from '../../../../hooks/index.js'
-import getChainBlockExplorerUrl from '../../../../utils/getChainBlockExplorerUrl.js'
-import { truncateAddress } from '../../../../utils/truncate.js'
 
 type DepositAddressValidatingStepProps = {
   txHashes: string[]
@@ -31,18 +28,6 @@ export const DepositAddressValidatingStep: FC<
         <Text style="subtitle2" css={{ mt: '4', mb: '2', textAlign: 'center' }}>
           Funds received. Your transaction is now in progress.
         </Text>
-        {status === 'delayed' ? (
-          <Flex
-            css={{ px: '4', py: '3', background: 'amber2', borderRadius: 12 }}
-          >
-            <Text style="subtitle2" css={{ textAlign: 'center' }}>
-              Your transaction is currently delayed. We apologize for the
-              inconvenience and appreciate your patience. After 5 minutes, your
-              transaction will either be processed or refunded. If this is
-              urgent, please contact support.
-            </Text>
-          </Flex>
-        ) : null}
         <Text
           color="subtle"
           style="body2"