+++ packages/ui/src/constants/walletCompatibility.ts
@@ -19,7 +19,10 @@ export const WalletChainIncompatible: Record<string, number[]> = {
 
 export const NormalizedWalletName: Record<string, string> = {
   phantomevm: 'phantom',
+  phantomsui: 'phantom',
+  phantombtc: 'phantom',
   magicedensol: 'magiceden',
   roninwallet: 'ronin',
-  portoevm: 'porto'
+  portoevm: 'porto',
+  surfwalletsui: 'surf'
 }