+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @reservoir0x/relay-kit-ui
 
+## 5.0.1
+
+### Patch Changes
+
+- e43a756: Improved copy for tx steps
+
+## 5.0.0
+
+### Major Changes
+
+- d38abdf: Revamped transaction steps
+
+### Patch Changes
+
+- 3bd044c: Fix duplicate deposit success event
+- Updated dependencies [d38abdf]
+  - @relayprotocol/relay-sdk@3.0.0
+  - @relayprotocol/relay-kit-hooks@2.0.6
+
 ## 4.0.19
 
 ### Patch Changes